Move Adpcm classes to their own namespace
