+++ src/DspAdpcm2/Formats/Adpcm/AdpcmChannelInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace DspAdpcm.Formats
+﻿namespace DspAdpcm.Formats.Adpcm
 {
     /// <summary>
     /// Defines the ADPCM information for a single
