+++ src/DspAdpcm2/Formats/Adpcm/AdpcmLoopContext.cs
@@ -1,4 +1,4 @@
-﻿namespace DspAdpcm.Formats
+﻿namespace DspAdpcm.Formats.Adpcm
 {
     internal class AdpcmLoopContext
     {
