+++ src/DspAdpcm2/Formats/AdpcmFormat.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using DspAdpcm.Codecs;
+using DspAdpcm.Formats.Adpcm;
 
 namespace DspAdpcm.Formats
 {
